LPS-108343 Rename folder from util to utils
