+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/components/Chart.es.js
@@ -22,7 +22,7 @@ import {
 	YAxis
 } from 'recharts';
 
-import {numberFormat} from '../util/numberFormat.es';
+import {numberFormat} from '../utils/numberFormat.es';
 
 const {useEffect, useMemo, useState} = React;
