+++ modules/dxp/apps/analytics/analytics-reports-web/test/js/utils/numberFormat.es.js
@@ -13,7 +13,7 @@ import {cleanup} from '@testing-library/react';
 
 import '@testing-library/jest-dom/extend-expect';
 
-import {numberFormat} from '../../../src/main/resources/META-INF/resources/js/util/numberFormat.es';
+import {numberFormat} from '../../../src/main/resources/META-INF/resources/js/utils/numberFormat.es';
 
 describe('numberFormat', () => {
 	afterEach(cleanup);
